added token configs * removed `api.token` in favor of `api.tokens` * added configs for tokens to override settings from main config
