+++ utils/config/token-config.go
@@ -1,0 +1,63 @@
+package config
+
+import (
+	"strconv"
+
+	log "github.com/codeshelldev/secured-signal-api/utils/logger"
+	"github.com/knadh/koanf/parsers/yaml"
+)
+
+type TOKEN_CONFIG_ struct {
+	TOKEN		string 		`koanf:"token"`
+	OVERRIDES 	*SETTING_	`koanf:"overrides"`
+}
+
+func LoadTokens() {
+	log.Debug("Loading Configs ", ENV.TOKENS_DIR)
+
+	LoadDir("tokens", ENV.TOKENS_DIR, tokensLayer, yaml.Parser())
+}
+
+func InitTokens() {
+	apiTokens := config.Strings("api.tokens")
+
+	var tokenConfigs []TOKEN_CONFIG_
+
+	tokensLayer.Unmarshal("tokens", &tokenConfigs)
+
+	overrides := ParseTokenConfigs(tokenConfigs)
+
+	for token, override := range overrides {
+		apiTokens = append(apiTokens, token)
+
+		ENV.SETTINGS[token] = override
+	}
+
+	if len(apiTokens) <= 0 {
+		log.Warn("No API TOKEN provided this is NOT recommended")
+
+		log.Info("Disabling Security Features due to incomplete Congfiguration")
+
+		ENV.INSECURE = true
+
+		// Set Blocked Endpoints on Config to User Layer Value
+		// => effectively ignoring Default Layer
+		config.Set("blockedendpoints", userLayer.Strings("blockeendpoints"))
+	}
+
+	if len(apiTokens) > 0 {
+		log.Debug("Registered " + strconv.Itoa(len(apiTokens)) + " Tokens")	
+
+		ENV.API_TOKENS = apiTokens
+	}
+}
+
+func ParseTokenConfigs(configs []TOKEN_CONFIG_) (map[string]*SETTING_) {
+	settings := map[string]*SETTING_{}
+
+	for _, config := range configs {
+		settings[config.TOKEN] = config.OVERRIDES
+	}
+
+	return settings
+}
